+++ test/hub_test.rb
@@ -7,8 +7,17 @@ class HubTest < Minitest::Test
   end
 
   def test_download_url_to_file
-    dst = "#{Dir.mktmpdir}/test.html"
-    assert_nil Torch::Hub.download_url_to_file("https://ankane.org/favicon.ico", dst)
+    dst = File.join(Dir.tmpdir, "tensor.pth")
+    assert_nil Torch::Hub.download_url_to_file(url, dst)
     assert File.exist?(dst)
+  end
+
+  def test_load_state_dict_from_url
+    state_dict = Torch::Hub.load_state_dict_from_url(url)
+    assert_equal Torch.load("test/support/tensor.pth").to_a, state_dict.to_a
+  end
+
+  def url
+    "https://github.com/ankane/torch.rb/raw/master/test/support/tensor.pth"
   end
 end